# openapi_azure_powerapps.json
{
  "swagger": "2.0",
  "info": {
    "title": "OSINTNeoAi Azure Cloud Intelligence API",
    "description": "Enterprise Microsoft Power Apps & Power Automate Custom Connector for OSINTNeoAi Master Hub, Tactical GIS Maps, and Forensic Intelligence Vault.",
    "version": "1.0.0"
  },
  "host": "osintneoai-app-949.azurewebsites.net",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/submit-victim": {
      "post": {
        "summary": "Submit Case Lead / Mutual Aid Intake",
        "description": "Submits a new investigation lead or intake report directly into the persistent forensic vault.",
        "operationId": "SubmitVictimReport",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "victim_name": { "type": "string", "example": "Jane Doe" },
                "contact_info": { "type": "string", "example": "jane@proton.me" },
                "incident_type": { "type": "string", "example": "Whistleblower Retaliation / Illegal Eviction" },
                "location": { "type": "string", "example": "Huntington Beach, CA" },
                "summary": { "type": "string", "example": "Witness tampering and procurement irregularities." }
              },
              "required": ["incident_type", "summary"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Submission confirmation"
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "summary": "Search Entities & APNs",
        "description": "Searches 17,000+ nodes, APN parcel records, and corporate entities.",
        "operationId": "SearchEntities",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Search keyword"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results list"
          }
        }
      }
    },
...
  }
}